Integration sync cleanup
