+++ src/lib/constants/integrations.ts
@@ -1,0 +1,1 @@
+export const DAILY_SYNC_COUNT_LIMIT = 5;
