+++ src/server/db/schema/index.ts
@@ -7,4 +7,4 @@ export * from './profiles.schema';
 export * from './subaccounts.schema';
 export * from './transactions.schema';
 export * from './integrations.schema';
-export * from './integration-to-subaccounts.schema';
+export * from './integration-links.schema';
