+++ src/server/db/schema/integration-links.schema.ts
@@ -1,14 +1,15 @@
 import { and, eq, relations, sql } from 'drizzle-orm';
-import { pgPolicy, pgTable, timestamp, uuid, primaryKey, text, integer } from 'drizzle-orm/pg-core';
+import { pgPolicy, pgTable, timestamp, uuid, text, integer } from 'drizzle-orm/pg-core';
 import { authenticatedRole } from 'drizzle-orm/supabase';
 
 import { integrations } from './integrations.schema';
 import { subaccounts } from './subaccounts.schema';
 import { authUid } from './utils';
 
-export const integrationToSubaccounts = pgTable(
-  'integration_to_subaccounts',
+export const integrationLinks = pgTable(
+  'integration_links',
   {
+    id: uuid().notNull().defaultRandom().primaryKey(),
     integrationId: uuid()
       .notNull()
       .references(() => integrations.id, { onDelete: 'cascade' }),
@@ -16,14 +17,13 @@ export const integrationToSubaccounts = pgTable(
       .notNull()
       .unique()
       .references(() => subaccounts.id, { onDelete: 'cascade' }),
-    integrationAccountId: text().unique().notNull(),
+    integrationAccountId: text().notNull(),
     lastSyncedAt: timestamp({ withTimezone: true }),
     syncCount: integer().notNull().default(0),
     createdAt: timestamp({ withTimezone: true }).notNull().defaultNow(),
   },
   (table) => [
-    primaryKey({ columns: [table.integrationId, table.subaccountId] }),
-    pgPolicy('integration_to_subaccounts_allow_all_for_owner_policy', {
+    pgPolicy('integration_links_allow_all_for_owner_policy', {
       as: 'permissive',
       for: 'all',
       to: authenticatedRole,
@@ -33,13 +33,13 @@ export const integrationToSubaccounts = pgTable(
   ],
 );
 
-export const integrationToSubaccountsRelations = relations(integrationToSubaccounts, ({ one }) => ({
+export const integrationLinksRelations = relations(integrationLinks, ({ one }) => ({
   integration: one(integrations, {
-    fields: [integrationToSubaccounts.integrationId],
+    fields: [integrationLinks.integrationId],
     references: [integrations.id],
   }),
   subaccount: one(subaccounts, {
-    fields: [integrationToSubaccounts.subaccountId],
+    fields: [integrationLinks.subaccountId],
     references: [subaccounts.id],
   }),
 }));
